cli: move file from get config file to shared
